fix: db model 수정 후 relationship 정의할 때 발생하는 에러 수정
